requirements.txt
streamlit
pymupdf
pillow
rapidfuzz
numpy
paddleocr
paddlepaddle
opencv-python-headless